src/handlers/messages/requirements.txt
python-telegram-bot
aws-lambda-powertools==1.28.0
openai==0.27.6
backoff==2.2.1
pydub==0.25.1
